# apps/api/src/@generated/core.api/models/Note.ts
/* tslint:disable */
/* eslint-disable */
/**
 * microservices/core
 * API for core microservice handling spaces and notes
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Note
 */
export interface Note {
    /**
     * 
     * @type {string}
     * @memberof Note
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Note
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof Note
     */
    body?: string;
    /**
     * 
     * @type {string}
     * @memberof Note
     */
    emoji?: string;
    /**
     * 
     * @type {string}
     * @memberof Note
     */
    type: NoteTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof Note
     */
    parentId?: string;
    /**
     * 
     * @type {string}
     * @memberof Note
     */
    sectionId?: string;
    /**
     * 
     * @type {string}
     * @memberof Note
     */
    spaceId: string;
    /**
     * 
     * @type {object}
     * @memberof Note
     */
    content?: object;
    /**
     * 
     * @type {Date}
     * @memberof Note
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Note
     */
    updatedAt: Date;
}


/**
 * @export
 */
export const NoteTypeEnum = {
    Note: 'note',
    Section: 'section',
    Paragraph: 'paragraph',
    NoteTask: 'note_task'
} as const;
export type NoteTypeEnum = typeof NoteTypeEnum[keyof typeof NoteTypeEnum];


/**
 * Check if a given object implements the Note interface.
 */
export function instanceOfNote(value: object): value is Note {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('spaceId' in value) || value['spaceId'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function NoteFromJSON(json: any): Note {
    return NoteFromJSONTyped(json, false);
}

export function NoteFromJSONTyped(json: any, ignoreDiscriminator: boolean): Note {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'],
        'body': json['body'] == null ? undefined : json['body'],
        'emoji': json['emoji'] == null ? undefined : json['emoji'],
        'type': json['type'],
        'parentId': json['parentId'] == null ? undefined : json['parentId'],
        'sectionId': json['sectionId'] == null ? undefined : json['sectionId'],
        'spaceId': json['spaceId'],
        'content': json['content'] == null ? undefined : json['content'],
        'createdAt': (new Date(json['createdAt'])),
        'updatedAt': (new Date(json['updatedAt'])),
    };
}

export function NoteToJSON(json: any): Note {
    return NoteToJSONTyped(json, false);
}

export function NoteToJSONTyped(value?: Note | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'body': value['body'],
        'emoji': value['emoji'],
        'type': value['type'],
        'parentId': value['parentId'],
        'sectionId': value['sectionId'],
        'spaceId': value['spaceId'],
        'content': value['content'],
        'createdAt': value['createdAt'].toISOString(),
        'updatedAt': value['updatedAt'].toISOString(),
    };
}