# apps/api/src/@generated/core.api/models/NoteCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * microservices/core
 * API for core microservice handling spaces and notes
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NoteCreate
 */
export interface NoteCreate {
    /**
     * 
     * @type {string}
     * @memberof NoteCreate
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof NoteCreate
     */
    body?: string;
    /**
     * 
     * @type {string}
     * @memberof NoteCreate
     */
    emoji?: string;
    /**
     * 
     * @type {string}
     * @memberof NoteCreate
     */
    type: NoteCreateTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof NoteCreate
     */
    parentId?: string;
    /**
     * 
     * @type {string}
     * @memberof NoteCreate
     */
    sectionId?: string;
    /**
     * 
     * @type {string}
     * @memberof NoteCreate
     */
    spaceId: string;
    /**
     * 
     * @type {object}
     * @memberof NoteCreate
     */
    content?: object;
}


/**
 * @export
 */
export const NoteCreateTypeEnum = {
    Note: 'note',
    Section: 'section',
    Paragraph: 'paragraph',
    NoteTask: 'note_task'
} as const;
export type NoteCreateTypeEnum = typeof NoteCreateTypeEnum[keyof typeof NoteCreateTypeEnum];


/**
 * Check if a given object implements the NoteCreate interface.
 */
export function instanceOfNoteCreate(value: object): value is NoteCreate {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('spaceId' in value) || value['spaceId'] === undefined) return false;
    return true;
}

export function NoteCreateFromJSON(json: any): NoteCreate {
    return NoteCreateFromJSONTyped(json, false);
}

export function NoteCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): NoteCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'body': json['body'] == null ? undefined : json['body'],
        'emoji': json['emoji'] == null ? undefined : json['emoji'],
        'type': json['type'],
        'parentId': json['parentId'] == null ? undefined : json['parentId'],
        'sectionId': json['sectionId'] == null ? undefined : json['sectionId'],
        'spaceId': json['spaceId'],
        'content': json['content'] == null ? undefined : json['content'],
    };
}

export function NoteCreateToJSON(json: any): NoteCreate {
    return NoteCreateToJSONTyped(json, false);
}

export function NoteCreateToJSONTyped(value?: NoteCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'body': value['body'],
        'emoji': value['emoji'],
        'type': value['type'],
        'parentId': value['parentId'],
        'sectionId': value['sectionId'],
        'spaceId': value['spaceId'],
        'content': value['content'],
    };
}