# apps/api/src/@generated/core.api/models/NoteUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * microservices/core
 * API for core microservice handling spaces and notes
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NoteUpdate
 */
export interface NoteUpdate {
    /**
     * 
     * @type {string}
     * @memberof NoteUpdate
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof NoteUpdate
     */
    body?: string;
    /**
     * 
     * @type {string}
     * @memberof NoteUpdate
     */
    emoji?: string;
    /**
     * 
     * @type {object}
     * @memberof NoteUpdate
     */
    content?: object;
}

/**
 * Check if a given object implements the NoteUpdate interface.
 */
export function instanceOfNoteUpdate(value: object): value is NoteUpdate {
    return true;
}

export function NoteUpdateFromJSON(json: any): NoteUpdate {
    return NoteUpdateFromJSONTyped(json, false);
}

export function NoteUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): NoteUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'body': json['body'] == null ? undefined : json['body'],
        'emoji': json['emoji'] == null ? undefined : json['emoji'],
        'content': json['content'] == null ? undefined : json['content'],
    };
}

export function NoteUpdateToJSON(json: any): NoteUpdate {
    return NoteUpdateToJSONTyped(json, false);
}

export function NoteUpdateToJSONTyped(value?: NoteUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'body': value['body'],
        'emoji': value['emoji'],
        'content': value['content'],
    };
}