# apps/api/src/@generated/core.api/models/Space.ts
/* tslint:disable */
/* eslint-disable */
/**
 * microservices/core
 * API for core microservice handling spaces and notes
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Space
 */
export interface Space {
    /**
     * 
     * @type {string}
     * @memberof Space
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Space
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Space
     */
    img?: string;
    /**
     * 
     * @type {Date}
     * @memberof Space
     */
    createdAt: Date;
}

/**
 * Check if a given object implements the Space interface.
 */
export function instanceOfSpace(value: object): value is Space {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    return true;
}

export function SpaceFromJSON(json: any): Space {
    return SpaceFromJSONTyped(json, false);
}

export function SpaceFromJSONTyped(json: any, ignoreDiscriminator: boolean): Space {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'img': json['img'] == null ? undefined : json['img'],
        'createdAt': (new Date(json['createdAt'])),
    };
}

export function SpaceToJSON(json: any): Space {
    return SpaceToJSONTyped(json, false);
}

export function SpaceToJSONTyped(value?: Space | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'img': value['img'],
        'createdAt': value['createdAt'].toISOString(),
    };
}