# apps/api/src/@generated/core.api/models/SpaceUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * microservices/core
 * API for core microservice handling spaces and notes
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SpaceUpdate
 */
export interface SpaceUpdate {
    /**
     * 
     * @type {string}
     * @memberof SpaceUpdate
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof SpaceUpdate
     */
    img?: string;
}

/**
 * Check if a given object implements the SpaceUpdate interface.
 */
export function instanceOfSpaceUpdate(value: object): value is SpaceUpdate {
    return true;
}

export function SpaceUpdateFromJSON(json: any): SpaceUpdate {
    return SpaceUpdateFromJSONTyped(json, false);
}

export function SpaceUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): SpaceUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'img': json['img'] == null ? undefined : json['img'],
    };
}

export function SpaceUpdateToJSON(json: any): SpaceUpdate {
    return SpaceUpdateToJSONTyped(json, false);
}

export function SpaceUpdateToJSONTyped(value?: SpaceUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'img': value['img'],
    };
}