# microservices/core/openapi.yaml
openapi: 3.0.3
info:
  title: microservices/core
  description: "API for core microservice handling spaces and notes"
  version: 0.0.1

tags:
  - name: space
    description: Рабочие пространства
  - name: note
    description: Заметки и контент

paths:
  /space:
    get:
      tags:
        - space
      summary: Получить список рабочих пространств
      operationId: getSpaces
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceList'
    post:
      tags:
        - space
      summary: Создать рабочее пространство
      operationId: createSpace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceCreate'
      responses:
        '201':
          description: Успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'

  /space/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: Уникальный идентификатор пространства
    get:
      tags:
        - space
      summary: Получить рабочее пространство
      operationId: getSpace
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
    put:
      tags:
        - space
      summary: Обновить рабочее пространство
      operationId: updateSpace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceUpdate'
      responses:
        '200':
          description: Успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
    delete:
      tags:
        - space
      summary: Удалить рабочее пространство
      operationId: deleteSpace
      responses:
        '204':
          description: Успешно удалено

  /note:
    get:
      tags:
        - note
      summary: Получить список заметок
      operationId: getNotes
      parameters:
        - in: query
          name: spaceId
          required: false
          schema:
            type: string
          description: ID пространства
        - in: query
          name: sectionId
          required: false
          schema:
            type: string
          description: ID раздела
        - in: query
          name: parentId
          required: false
          schema:
            type: string
          description: ID родителя
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteList'
    post:
      tags:
        - note
      summary: Создать заметку
      operationId: createNote
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteCreate'
      responses:
        '201':
          description: Успешно создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'

  /note/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: Уникальный идентификатор заметки
    get:
      tags:
        - note
      summary: Получить заметку
      operationId: getNote
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
    put:
      tags:
        - note
      summary: Обновить заметку
      operationId: updateNote
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteUpdate'
      responses:
        '200':
          description: Успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
    delete:
      tags:
        - note
      summary: Удалить заметку
      operationId: deleteNote
      responses:
        '204':
          description: Успешно удалено

components:
  schemas:
    SpaceList:
      type: array
      items:
        $ref: '#/components/schemas/Space'

    Space:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        img:
          type: string
        createdAt:
          type: string
          format: date-time

    SpaceCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        img:
          type: string

    SpaceUpdate:
      type: object
      properties:
        name:
          type: string
        img:
          type: string

    NoteList:
      type: array
      items:
        $ref: '#/components/schemas/Note'

    Note:
      type: object
      required: [id, title, type, spaceId, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        emoji:
          type: string
        type:
          type: string
          enum: [note, section, paragraph, note_task]
        parentId:
          type: string
        sectionId:
          type: string
        spaceId:
          type: string
        content:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    NoteCreate:
      type: object
      required: [title, type, spaceId]
      properties:
        title:
          type: string
        body:
          type: string
        emoji:
          type: string
        type:
          type: string
          enum: [note, section, paragraph, note_task]
        parentId:
          type: string
        sectionId:
          type: string
        spaceId:
          type: string
        content:
          type: object

    NoteUpdate:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        emoji:
          type: string
        content:
          type: object